fix: Move gatekeeper reporting to 'status' cli command
